+++ FBSDKCoreKit/FBSDKCoreKit/AppEvents/Internal/ML/FBSDKModelManager.h
@@ -24,13 +24,17 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef void (^FBSDKDownloadCompletionBlock)(BOOL success);
+typedef void (^FBSDKDownloadCompletionBlock)(void);
 
 @interface FBSDKModelManager : NSObject
 
 + (void)enable;
-+ (nullable NSDictionary *)getRulesForKey:(NSString *)useCaseKey;
-+ (nullable NSString *)getWeightsPath:(NSString *)useCaseKey;
++ (nullable NSDictionary *)getRulesForKey:(NSString *)useCase;
++ (nullable NSData *)getWeightsForKey:(NSString *)useCase;
++ (nullable NSArray *)getThresholdsForKey:(NSString *)useCase;
+
++ (BOOL)processIntegrity:(nullable NSString *)param;
++ (NSString *)processSuggestedEvents:(NSString *)textFeature denseData:(nullable float *)denseData;
 
 @end