+++ FBSDKCoreKit/FBSDKCoreKit/AppEvents/Internal/ViewHierarchy/FBSDKViewHierarchy.m
@@ -150,7 +150,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   return children;
 }
 
-+ (nullable NSObject *)getParent:(NSObject *)obj
++ (nullable NSObject *)getParent:(nullable NSObject *)obj
 {
   if ([obj isKindOfClass:[UIView class]]) {
     UIView *superview = ((UIView *)obj).superview;
@@ -225,14 +225,14 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   return [NSArray arrayWithArray:path];
 }
 
-+ (NSDictionary<NSString *, id> *)getAttributesOf:(NSObject *)obj parent:(NSObject *)parent
++ (NSDictionary<NSString *, id> *)getAttributesOf:(NSObject *)obj parent:(NSObject * _Nullable)parent
 {
   NSMutableDictionary *componentInfo = [NSMutableDictionary dictionary];
   componentInfo[CODELESS_MAPPING_CLASS_NAME_KEY] = NSStringFromClass([obj class]);
 
   if (![FBSDKViewHierarchy isUserInputView:obj]) {
     NSString *text = [FBSDKViewHierarchy getText:obj];
-    if (text) {
+    if (text.length > 0) {
       componentInfo[CODELESS_MAPPING_TEXT_KEY] = text;
     }
   } else {
@@ -241,7 +241,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   }
 
   NSString *hint = [FBSDKViewHierarchy getHint:obj];
-  if (hint) {
+  if (hint.length > 0) {
     componentInfo[CODELESS_MAPPING_HINT_KEY] = hint;
   }
 
@@ -336,7 +336,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   return indexPath;
 }
 
-+ (nullable NSString *)getText:(NSObject *)obj
++ (NSString *)getText:(nullable NSObject *)obj
 {
   NSString *text = nil;
 
@@ -352,6 +352,10 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
     NSMutableArray *titles = [NSMutableArray array];
 
     for (NSInteger i = 0; i < sections; i++) {
+      NSInteger numberOfRow = [picker numberOfRowsInComponent:i];
+      if (numberOfRow <= 0) {
+        continue;
+      }
       NSInteger row = [picker selectedRowInComponent:i];
       NSString *title;
       if ([picker.delegate
@@ -388,7 +392,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
     text = attributedText.string;
   }
 
-  return text.length > 0 ? text : nil;
+  return text ?: @"";
 }
 
 + (nullable NSDictionary<NSString *, id> *)getTextStyle:(NSObject *)obj
@@ -420,12 +424,14 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   return nil;
 }
 
-+ (nullable NSString *)getHint:(NSObject *)obj
++ (NSString *)getHint:(nullable NSObject *)obj
 {
   NSString *hint = nil;
 
   if ([obj isKindOfClass:[UITextField class]]) {
-    hint = ((UITextField *)obj).placeholder;
+    UITextField *textField = (UITextField *)obj;
+    hint = textField.placeholder ?: @"";
+    hint = [hint stringByAppendingString:[self recursiveGetLabelsFromView:textField]];
   } else if ([obj isKindOfClass:[UINavigationController class]]) {
     UIViewController *top = ((UINavigationController *)obj).topViewController;
     if (top) {
@@ -433,7 +439,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
     }
   }
 
-  return hint.length > 0 ? hint : nil;
+  return hint ?: @"";
 }
 
 + (NSUInteger)getClassBitmask:(NSObject *)obj
@@ -495,7 +501,7 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   }
 
   NSString *text = [FBSDKViewHierarchy getText:obj];
-  return text && [FBSDKAppEventsUtility isSensitiveUserData:text];
+  return text.length > 0 && [FBSDKAppEventsUtility isSensitiveUserData:text];
 }
 
 + (nullable NSDictionary<NSString *, id> *)recursiveCaptureTreeWithCurrentNode:(NSObject *)currentNode
@@ -642,7 +648,10 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
   if ([obj isKindOfClass:[UIView class]]) {
     return ((UIView *)obj).tag;
   } else if ([obj isKindOfClass:[UIViewController class]]) {
-    return ((UIViewController *)obj).view.tag;
+    UIViewController *vc = (UIViewController *)obj;
+    if (vc.isViewLoaded) {
+      return ((UIViewController *)obj).view.tag;
+    }
   }
 
   return 0;
@@ -673,6 +682,18 @@ void fb_dispatch_on_default_thread(dispatch_block_t block) {
            CODELESS_VIEW_TREE_OFFSET_Y_KEY: @((int)offset.y),
            CODELESS_VIEW_TREE_VISIBILITY_KEY: view.isHidden ? @4 : @0
            };
+}
+
++ (NSString *)recursiveGetLabelsFromView:(UIView *)view
+{
+  NSString *str = @"";
+  for (UIView *subview in view.subviews) {
+    str = [str stringByAppendingString:[self recursiveGetLabelsFromView:subview]];
+  }
+  if ([view isKindOfClass:[UILabel class]] && ((UILabel *)view).text.length > 0) {
+    str = [str stringByAppendingFormat:@" %@", ((UILabel *)view).text];
+  }
+  return str;
 }
 
 @end