+++ FBSDKCoreKit/FBSDKCoreKit/Internal/FBSDKFeatureManager.m
@@ -46,13 +46,6 @@ NS_ASSUME_NONNULL_BEGIN
   }];
 }
 
-+ (void)disableFeature:(NSString *)featureName
-{
-  [[NSUserDefaults standardUserDefaults] setObject:[FBSDKSettings sdkVersion] forKey:[FBSDKFeatureManagerPrefix stringByAppendingString:featureName]];
-}
-
-#pragma mark - Private methods
-
 + (BOOL)isEnabled:(FBSDKFeature)feature
 {
   if (FBSDKFeatureCore == feature) {
@@ -66,6 +59,13 @@ NS_ASSUME_NONNULL_BEGIN
     return [FBSDKFeatureManager isEnabled:parentFeature] && [self checkGK:feature];
   }
 }
+
++ (void)disableFeature:(NSString *)featureName
+{
+  [[NSUserDefaults standardUserDefaults] setObject:[FBSDKSettings sdkVersion] forKey:[FBSDKFeatureManagerPrefix stringByAppendingString:featureName]];
+}
+
+#pragma mark - Private methods
 
 + (FBSDKFeature)getParentFeature:(FBSDKFeature)feature
 {
@@ -98,13 +98,14 @@ NS_ASSUME_NONNULL_BEGIN
     case FBSDKFeatureAAM: featureName = @"AAM"; break;
     case FBSDKFeaturePrivacyProtection: featureName = @"PrivacyProtection"; break;
     case FBSDKFeatureSuggestedEvents: featureName = @"SuggestedEvents"; break;
-    case FBSDKFeaturePIIFiltering: featureName = @"PIIFiltering"; break;
-    case FBSDKFeatureMTML: featureName = @"MTML"; break;
+    case FBSDKFeatureIntelligentIntegrity: featureName = @"IntelligentIntegrity"; break;
+    case FBSDKFeatureModelRequest: featureName = @"ModelRequest"; break;
     case FBSDKFeatureEventDeactivation: featureName = @"EventDeactivation"; break;
     case FBSDKFeatureInstrument: featureName = @"Instrument"; break;
     case FBSDKFeatureCrashReport: featureName = @"CrashReport"; break;
     case FBSDKFeatureCrashShield: featureName = @"CrashShield"; break;
     case FBSDKFeatureErrorReport: featureName = @"ErrorReport"; break;
+    case FBSDKFeatureMonitoring: featureName = @"Monitoring"; break;
 
     case FBSDKFeatureLogin: featureName = @"LoginKit"; break;
 
@@ -128,8 +129,9 @@ NS_ASSUME_NONNULL_BEGIN
     case FBSDKFeatureAAM:
     case FBSDKFeaturePrivacyProtection:
     case FBSDKFeatureSuggestedEvents:
-    case FBSDKFeaturePIIFiltering:
-    case FBSDKFeatureMTML:
+    case FBSDKFeatureIntelligentIntegrity:
+    case FBSDKFeatureModelRequest:
+    case FBSDKFeatureMonitoring:
       return NO;
     case FBSDKFeatureLogin:
     case FBDSDKFeatureShare: