+++ FBSDKGamingServicesKit/FBSDKGamingServicesKit/FBSDKGamingImageUploader.h
@@ -38,8 +38,36 @@ NS_SWIFT_NAME(GamingImageUploader)
  @param completionHandler a callback that is fired dependent on the configuration.
   Fired when the upload completes or when the users returns to the caller app
   after the media dialog is shown.
+
+@deprecated - This method is deprecated. Please use uploadImageWithConfiguration:andResultCompletionHandler: instead.
+*/
++ (void)uploadImageWithConfiguration:(FBSDKGamingImageUploaderConfiguration * _Nonnull)configuration
+                andCompletionHandler:(FBSDKGamingServiceCompletionHandler _Nonnull)completionHandler
+DEPRECATED_MSG_ATTRIBUTE("Please use uploadImageWithConfiguration:andResultCompletionHandler: instead");
+
+/**
+Runs an upload to a users Gaming Media Library with the given configuration
+
+@param configuration model object contain the content that will be uploaded
+@param completionHandler a callback that is fired dependent on the configuration.
+ Fired when the upload completes or when the users returns to the caller app
+ after the media dialog is shown.
  */
 + (void)uploadImageWithConfiguration:(FBSDKGamingImageUploaderConfiguration * _Nonnull)configuration
-                andCompletionHandler:(FBSDKGamingServiceCompletionHandler _Nonnull)completionHandler;
+          andResultCompletionHandler:(FBSDKGamingServiceResultCompletionHandler _Nonnull)completionHandler;
+
+/**
+Runs an upload to a users Gaming Media Library with the given configuration
+
+@param configuration model object contain the content that will be uploaded
+@param completionHandler a callback that is fired dependent on the configuration.
+ Fired when the upload completes or when the users returns to the caller app
+ after the media dialog is shown.
+@param progressHandler an optional callback that is fired multiple times as
+ bytes are transferred to Facebook.
+*/
++ (void)uploadImageWithConfiguration:(FBSDKGamingImageUploaderConfiguration * _Nonnull)configuration
+                   completionHandler:(FBSDKGamingServiceResultCompletionHandler _Nonnull)completionHandler
+                  andProgressHandler:(FBSDKGamingServiceProgressHandler _Nullable)progressHandler;
 
 @end