+++ FBSDKGamingServicesKit/FBSDKGamingServicesKit/FBSDKGamingImageUploaderConfiguration.m
@@ -20,9 +20,14 @@
 
 @implementation FBSDKGamingImageUploaderConfiguration
 
+- (instancetype)init
+{
+  return [super init];
+}
+
 - (instancetype)initWithImage:(UIImage * _Nonnull)image
-                caption:(NSString * _Nullable)caption
-shouldLaunchMediaDialog:(BOOL)shouldLaunchMediaDialog
+                      caption:(NSString * _Nullable)caption
+      shouldLaunchMediaDialog:(BOOL)shouldLaunchMediaDialog
 {
   if (self = [super init]) {
     _image = image;