+++ FBSDKGamingServicesKit/FBSDKGamingServicesKit/FBSDKGamingServiceCompletionHandler.h
@@ -17,6 +17,33 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #import <Foundation/Foundation.h>
+#import <AvailabilityMacros.h>
 
+/**
+ Main completion handling of any Gaming Service (Friend Finder, Image/Video Upload).
+
+ @param success whether the call to the service was considered a success.
+ @param error the error that occured during the service call, if any.
+ */
 typedef void (^FBSDKGamingServiceCompletionHandler)(BOOL success, NSError * _Nullable error)
 NS_SWIFT_NAME(GamingServiceCompletionHandler);
+
+/**
+Main completion handling of any Gaming Service (Friend Finder, Image/Video Upload).
+
+@param success whether the call to the service was considered a success.
+@param result the result that was returned by the service, if any.
+@param error the error that occured during the service call, if any.
+*/
+typedef void (^FBSDKGamingServiceResultCompletionHandler)(BOOL success, NSString * _Nullable result, NSError * _Nullable error)
+NS_SWIFT_NAME(GamingServiceCompletionHandler);
+
+/**
+Main completion handling of any Gaming Service (Friend Finder, Image/Video Upload).
+
+@param bytesSent the number of bytes sent since the last invocation
+@param totalBytesSent the total number of bytes sent
+@param totalBytesExpectedToSend the number of bytes that remain to be sent
+*/
+typedef void (^FBSDKGamingServiceProgressHandler)(int64_t bytesSent, int64_t totalBytesSent, int64_t totalBytesExpectedToSend)
+NS_SWIFT_NAME(GamingServiceProgressHandler);