+++ FBSDKGamingServicesKit/FBSDKGamingServicesKit/FBSDKGamingServicesKit.h
@@ -22,6 +22,8 @@
 #import <FBSDKGamingServicesKit/FBSDKGamingImageUploader.h>
 #import <FBSDKGamingServicesKit/FBSDKGamingImageUploaderConfiguration.h>
 #import <FBSDKGamingServicesKit/FBSDKGamingServiceCompletionHandler.h>
+#import <FBSDKGamingServicesKit/FBSDKGamingVideoUploader.h>
+#import <FBSDKGamingServicesKit/FBSDKGamingVideoUploaderConfiguration.h>
 
 #else
 
@@ -29,6 +31,8 @@
 #import "FBSDKGamingImageUploader.h"
 #import "FBSDKGamingImageUploaderConfiguration.h"
 #import "FBSDKGamingServiceCompletionHandler.h"
+#import "FBSDKGamingVideoUploader.h"
+#import "FBSDKGamingVideoUploaderConfiguration.h"
 
 #endif