+++ FBSDKGamingServicesKit/FBSDKGamingServicesKit/Internal/FBSDKGamingServiceController.m
@@ -49,15 +49,18 @@ static NSURL* FBSDKGamingServicesUrl(FBSDKGamingServiceType serviceType, NSStrin
 @implementation FBSDKGamingServiceController
 {
   FBSDKGamingServiceType _serviceType;
-  FBSDKGamingServiceCompletionHandler _completionHandler;
+  FBSDKGamingServiceResultCompletionHandler _completionHandler;
+  id _pendingResult;
 }
 
 - (instancetype)initWithServiceType:(FBSDKGamingServiceType)serviceType
-                  completionHandler:(FBSDKGamingServiceCompletionHandler)completionHandler
+                  completionHandler:(FBSDKGamingServiceResultCompletionHandler)completionHandler
+                      pendingResult:(id)pendingResult
 {
   if (self = [super init]) {
     _serviceType = serviceType;
     _completionHandler = completionHandler;
+    _pendingResult = pendingResult;
   }
   return self;
 }
@@ -82,14 +85,18 @@ static NSURL* FBSDKGamingServicesUrl(FBSDKGamingServiceType serviceType, NSStrin
   }
 
   if (error) {
-    _completionHandler(false, [FBSDKError
-                               errorWithCode:FBSDKErrorBridgeAPIInterruption
-                               message:@"Error occured while interacting with Gaming Services"
-                               underlyingError:error]);
+    _completionHandler(false,
+                       nil,
+                       [FBSDKError
+                       errorWithCode:FBSDKErrorBridgeAPIInterruption
+                       message:@"Error occured while interacting with Gaming Services"
+                       underlyingError:error]);
   } else {
-    _completionHandler(false, [FBSDKError
-                               errorWithCode:FBSDKErrorBridgeAPIInterruption
-                               message:@"An Unknown error occured while interacting with Gaming Services"]);
+    _completionHandler(false,
+                       nil,
+                       [FBSDKError
+                        errorWithCode:FBSDKErrorBridgeAPIInterruption
+                        message:@"An Unknown error occured while interacting with Gaming Services"]);
   }
 
   _completionHandler = nil;
@@ -97,7 +104,7 @@ static NSURL* FBSDKGamingServicesUrl(FBSDKGamingServiceType serviceType, NSStrin
 
 - (void)completeSuccessfully
 {
-  _completionHandler(true, nil);
+  _completionHandler(true, _pendingResult, nil);
   _completionHandler = nil;
 }